+++ system/library/Controller/ControlPanel.php
@@ -12,70 +12,6 @@ class Controller_ControlPanel
 {
 
     /**
-     * Returns list of control panel logins
-     * @url /api/controlpanel/:website_id
-     * @method GET
-     */
-    public static function listAction($website_id)
-    {
-        $app = Slim::getInstance();
-        $response = $app->response();
-        $response->header('Content-Type', 'application/json');
-        try {
-            $mgr = new Model_Website();
-            $website = $mgr->websiteExists($website_id);
-            if (!$website) {
-                $response->status(404);
-                $response->body(json_encode(array('success' => false, 'message' => "Website not found")));
-                return $response;
-            }
-            $mgr = new Model_ControlPanel();
-            $results = $mgr->getControlPanelLogins($website_id);
-            $response->body(json_encode(array('success' => true, 'records' => $results)));
-            return $response;
-        } catch (Exception $e) {
-            $app->getLog()->error("Error listing control panels for website " . $website_id . ". - " . $e->getMessage());
-            $response->status(500);
-            $response->body(json_encode(array('success' => false, 'message' => $e->getMessage())));
-            return $response;
-        }
-    }
-
-    /**
-     * Returns details for an control panel login
-     * @url /api/controlpanel/:website_id/:id
-     * @method GET
-     */
-    public static function detailsAction($website_id, $id)
-    {
-        $app = Slim::getInstance();
-        $response = $app->response();
-        $response->header('Content-Type', 'application/json');
-        try {
-            $mgr = new Model_Website();
-            $website = $mgr->websiteExists($website_id);
-            if (!$website) {
-                $response->status(404);
-                $response->body(json_encode(array('success' => false, 'message' => "Website not found")));
-                return $response;
-            }
-            $mgr = new Model_ControlPanel();
-            $results = $mgr->getControlPanelDetails($id, $website_id);
-            if (!$results) {
-                $response->status(404);
-                $response->body(json_encode(array('success' => false, 'message' => "Control panel login credentials not found")));
-                return $response;
-            }
-            return $response->body(json_encode(array('success' => true, 'record' => $results)));
-        } catch (Exception $e) {
-            $app->getLog()->error("Error showing control panel {$id} for website " . $website_id . ". - " . $e->getMessage());
-            $response->status(500);
-            $response->body(json_encode(array('success' => false, 'message' => $e->getMessage())));
-            return $response;
-        }
-    }
-
-    /**
      * Adds a new control panel login
      * @url /api/controlpanel/:website_id
      * @method POST
@@ -84,33 +20,30 @@ class Controller_ControlPanel
     {
         $app = Slim::getInstance();
         $response = $app->response();
-        $response->header('Content-Type', 'application/json');
-        try {
-            $mgr = new Model_Website();
-            $website = $mgr->websiteExists($website_id);
-            if (!$website) {
-                $response->status(404);
-                $response->body(json_encode(array('success' => false, 'message' => "Website not found")));
-                return $response;
-            }
-            $mgr = new Model_ControlPanel();
-            $results = $mgr->addControlPanelLogin($website_id, $app->request()->post());
-            $app->getLog()->info("Control panel {$results['id']} added for website " . $website_id . ".");
-            $response->status(201);
-            $response->body(json_encode(array('success' => true, 'message' => 'Control panel Login has been added.', 'record' => $results)));
+        $mgr = new Model_Website();
+        $website = $mgr->getWebsite($website_id, false);
+        if (!$website) {
+            $app->render('error/not-found.twig', array('message' => 'The website you requested does not exist.'), 404);
             return $response;
-        } catch (Exception $e) {
-            if ($e instanceof Validate_Exception) {
-                $response->status(400);
-                $response->body(json_encode(array('success' => false, 'message' => $e->getMessage(), 'errors' => $e->getErrors()->getErrors())));
+        }
+        $results = Model_ControlPanel::$default;
+        if ($app->request()->isPost()) {
+            try {
+                $results = array_merge($results, array_intersect_key($app->request()->post('controlpanel'), $results));
+                $mgr = new Model_ControlPanel();
+                $results = $mgr->addControlPanelLogin($website_id, $results);
+                $app->flash('success', "Control panel {$results['id']} added for website {$website['name']}.");
+                $response->redirect("/website/{$website['id']}");
                 return $response;
-            } else {
-                $app->getLog()->error("Error adding control panel for website " . $website_id . ". - " . $e->getMessage());
-                $response->status(500);
-                $response->body(json_encode(array('success' => false, 'message' => $e->getMessage())));
+            } catch (Validate_Exception $e) {
+                $app->render('controlpanel/add.twig', array('website' => $website, 'results' => $results, 'errors' => $e->getErrors()));
+                return $response;
+            } catch (Exception $e) {
+                $app->render('error/error.twig', array('exception' => $e, 'message' => 'Error adding control panel.'), 500);
                 return $response;
             }
         }
+        $app->render('controlpanel/add.twig', array('website' => $website, 'results' => $results));
     }
 
     /**
@@ -122,40 +55,34 @@ class Controller_ControlPanel
     {
         $app = Slim::getInstance();
         $response = $app->response();
-        $response->header('Content-Type', 'application/json');
-        try {
-            $mgr = new Model_Website();
-            $website = $mgr->getWebsite($website_id);
-            if (!$website) {
-                $response->status(404);
-                $response->body(json_encode(array('success' => false, 'message' => "Website not found.")));
-                return $response;
-            }
-            $mgr = new Model_ControlPanel();
-            $results = $mgr->getControlPanelDetails($id, $website_id);
-            $app->getLog()->info("Control panel {$id} updated for website " . $website_id . ".");
-            if (!$results) {
-                $response->status(404);
-                $response->body(json_encode(array('success' => false, 'message' => "Control panel login credentials not found")));
-                return $response;
-            }
-            $results = array_merge($results, array_intersect_key($app->request()->post(), $results));
-            $results = $mgr->updateControlPanelLogin($id, $results, $website_id);
-            $response->status(200);
-            $response->body(json_encode(array('success' => true, 'message' => 'Control panel login has been updated.', 'record' => $results)));
+        $mgr = new Model_Website();
+        $website = $mgr->getWebsite($website_id);
+        if (!$website) {
+            $app->render('error/not-found.twig', array('message' => 'The website you requested does not exist.'), 404);
             return $response;
-        } catch (Exception $e) {
-            if ($e instanceof Validate_Exception) {
-                $response->status(400);
-                $response->body(json_encode(array('success' => false, 'message' => $e->getMessage(), 'errors' => $e->getErrors()->getErrors())));
+        }
+        $mgr = new Model_ControlPanel();
+        $results = $mgr->getControlPanelDetails($id, $website_id);
+        if (!$results) {
+            $app->render('error/not-found.twig', array('message' => 'Control panel credentials not found.'), 404);
+            return $response;
+        }
+        if ($app->request()->isPost() || $app->request()->isPut()) {
+            try {
+                $results = array_merge($results, array_intersect_key($app->request()->post('controlpanel'), $results));
+                $results = $mgr->updateControlPanelLogin($id, $results, $website_id);
+                $app->flash('success', "Control panel credentials {$id} have been updated for website {$website['name']}.");
+                $response->redirect("/website/{$website['id']}");
                 return $response;
-            } else {
-                $app->getLog()->error("Error updating control panel {$id} for website " . $website_id . ". - " . $e->getMessage());
-                $response->status(500);
-                $response->body(json_encode(array('success' => false, 'message' => $e->getMessage())));
+            } catch (Validate_Exception $e) {
+                $app->render('controlpanel/edit.twig', array('website' => $website, 'results' => $results, 'errors' => $e->getErrors()));
+                return $response;
+            } catch (Exception $e) {
+                $app->render('error/error.twig', array('exception' => $e, 'message' => 'Error updating control panel.'), 500);
                 return $response;
             }
         }
+        $app->render('controlpanel/edit.twig', array('website' => $website, 'results' => $results));
     }
 
     /**
@@ -167,38 +94,32 @@ class Controller_ControlPanel
     {
         $app = Slim::getInstance();
         $response = $app->response();
-        $response->header('Content-Type', 'application/json');
-        try {
-            $mgr = new Model_Website();
-            $website = $mgr->getWebsite($website_id);
-            if (!$website) {
-                $response->status(404);
-                $response->body(json_encode(array('success' => false, 'message' => "Website not found.")));
-                return $response;
-            }
-            $mgr = new Model_ControlPanel();
-            $results = $mgr->getControlPanelDetails($id, $website_id);
-            if (!$results) {
-                $response->status(404);
-                $response->body(json_encode(array('success' => false, 'message' => "Control panel login credentials not found")));
-                return $response;
-            }
-            $results = $mgr->deleteControlPanelLogin($id);
-            $app->getLog()->info("Control panel {$id} deleted from website " . $website_id . ".");
-            $response->status(204);
-            $response->body(json_encode(array('success' => true, 'message' => 'Control panel login has been deleted.')));
+        $mgr = new Model_Website();
+        $website = $mgr->getWebsite($website_id);
+        if (!$website) {
+            $app->render('error/not-found.twig', array('message' => 'The website you requested does not exist.'), 404);
             return $response;
-        } catch (Exception $e) {
-            if ($e instanceof Validate_Exception) {
-                $response->status(400);
-                $response->body(json_encode(array('success' => false, 'message' => $e->getMessage(), 'errors' => $e->getErrors()->getErrors())));
+        }
+        $mgr = new Model_ControlPanel();
+        $results = $mgr->getControlPanelDetails($id, $website_id);
+        if (!$results) {
+            $app->render('error/not-found.twig', array('message' => 'Control panel credentials not found.'), 404);
+            return $response;
+        }
+        if ($app->request()->isPost() || $app->request()->isDelete()) {
+            try {
+                $results = $mgr->deleteControlPanelLogin($id);
+                $app->flash('info', "Control panel {$id} deleted from website {$website['name']}.");
+                $response->redirect("/website/{$website_id}");
                 return $response;
-            } else {
-                $app->getLog()->error("Error deleting control panel {$id} for website " . $website_id . ". - " . $e->getMessage());
-                $response->status(500);
-                $response->body(json_encode(array('success' => false, 'message' => $e->getMessage())));
+            } catch (Validate_Exception $e) {
+                $app->render('controlpanel/delete.twig', array('website' => $website, 'errors' => $e->getErrors()));
+                return $response;
+            } catch (Exception $e) {
+                $app->render('error/error.twig', array('exception' => $e, 'message' => "Error deleting control panel {$id} for website {$website['name']}."), 500);
                 return $response;
             }
         }
+        $app->render('controlpanel/delete.twig', array('website' => $website, 'results' => $results));
     }
 }
